+++ app/src/main/java/com/juvinal/pay/ui/screens/inApp/dashboard/HomeScreenViewModel.kt
@@ -62,7 +62,7 @@ class HomeScreenViewModel(
                             accountSavings = response.body()?.data?.accountSavings!!.toDouble(),
                             loanBalance = response.body()?.data?.loanBalance!!,
                             guaranteedAmounts = response.body()?.data?.guaranteedAmounts!!,
-                            netSavings = response.body()?.data?.netSavings!!,
+                            netSavings = response.body()?.data?.netSavings!!.toDouble(),
                             accountShareCapital = response.body()?.data?.accountShareCapital!!.toDouble(),
                             loanAmountQualified = response.body()?.data?.loanAmountQualified!!.toDouble(),
                             loadingStatus = LoadingStatus.SUCCESS
